[FE] [ADD] add search bar for restaurants (#33)
* [RESTO] [FIX] add url checker before usage, this fix axios problem (dont ask why, I also dont know why...)

* [RESTO] [ADD] add filter for restoraunts

* [RESTO] [ADD] use i18n

* [RESTO] [FIX] fix smaller issues with coding style and change wording from Location to Ort

+++ restoMobile/src/services/restoCalls.ts
@@ -1,13 +1,15 @@
 import axios from 'axios';
 // @ts-ignore
 import { API_URL } from '@env';
-import { faA } from '@fortawesome/free-solid-svg-icons';
 
 const baseURL = API_URL;
 const menuDesignUrl = `${API_URL}menuDesigns/`;
 
 export const getAllResto = async () => {
   try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
     const response = await axios({
       method: 'GET',
       url: baseURL + 'restaurants/',
@@ -21,6 +23,9 @@ export const getAllResto = async () => {
 
 export const getAllRestaurantsByUser = async (body: any) => {
   try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
     const response = await axios({
       method: "GET",
       url: baseURL + 'restaurants/user/resto',
@@ -36,8 +41,31 @@ export const getAllRestaurantsByUser = async (body: any) => {
   }
 };
 
+export const getAllRestaurantsByUserAndFilter = async (token: string, filter: string) => {
+  try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
+    const response = await axios({
+      method: "POST",
+      url: baseURL + 'search/restaurants/',
+      data: { "token": token,"filter": filter },
+      headers: {
+        "content-type": "application/json",
+      },
+    });
+    return response.data;
+  } catch (error) {
+    console.error("Error fetching all restaurants:", error);
+    throw new Error("Failed to fetch all restaurants");
+  }
+};
+
 export const getRestaurantByName = async(name: string) => {
   try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
     const response = await axios({
       method: 'GET',
       url: baseURL + 'restaurants/'+name,
@@ -54,6 +82,9 @@ export const getRestaurantByName = async(name: string) => {
 
 export const addRestaurant = async(restaurantData: any) => {
   try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
     const response = await axios({
       method: 'POST',
       url: baseURL + 'restaurants/',
@@ -71,6 +102,9 @@ export const addRestaurant = async(restaurantData: any) => {
 
 export const deleteRestaurantByName = async (restaurantName: string) => {
   try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
     await axios({
       method: 'DELETE',
       url: baseURL +'restaurants/' + restaurantName,
@@ -83,6 +117,9 @@ export const deleteRestaurantByName = async (restaurantName: string) => {
 
 export const getAllMenuDesigns = async () => {
   try {
+    if (menuDesignUrl === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
     const response = await axios({
       method: "GET",
       url: menuDesignUrl
@@ -96,6 +133,9 @@ export const getAllMenuDesigns = async () => {
 
 export const editResto = async (restoName: string, body: any) => {
   try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
     const response = await axios({
       url: baseURL + 'restaurants/' + restoName,
       method: "PUT",
@@ -118,6 +158,9 @@ export const editResto = async (restoName: string, body: any) => {
 
 export const getRestoByName = async (restoName: string) => {
   try {
+    if (baseURL === undefined) {
+      throw new Error("baseUrl is not defined");
+    }
     const response = await axios({
       url: baseURL + 'restaurants/' + restoName,
       method: "GET",
